feat: reorganise file structure
