+++ backend/question-service/server/app/utils/exception_handler.py
@@ -1,0 +1,7 @@
+from fastapi import HTTPException
+from utils.jsend import jsend_response, JSendStatus
+
+
+def custom_http_exception_handler(request, exc: HTTPException):
+    """Custom exception handler to return JSend-style responses for HTTPExceptions"""
+    return jsend_response(JSendStatus.ERROR, code=exc.status_code, message=exc.detail)
